+++ lib/AL_USDMaya/AL/usdmaya/nodes/ProxyDrawOverride.cpp
@@ -59,6 +59,7 @@ public:
 //----------------------------------------------------------------------------------------------------------------------
 MString ProxyDrawOverride::kDrawDbClassification("drawdb/geometry/AL_usdmaya");
 MString ProxyDrawOverride::kDrawRegistrantId("pxrUsd");
+MUint64 ProxyDrawOverride::s_lastRefreshFrameStamp = 0;
 
 //----------------------------------------------------------------------------------------------------------------------
 ProxyDrawOverride::ProxyDrawOverride(const MObject& obj)
@@ -135,8 +136,6 @@ MUserData* ProxyDrawOverride::prepareForDraw(
     return NULL;
   }
 
-  data->m_params.showGuides = data->m_shape->displayGuidesPlug().asBool();
-  data->m_params.showRender = data->m_shape->displayRenderGuidesPlug().asBool();
   data->m_rootPrim = data->m_shape->getRootPrim();
   data->m_engine = engine;
 
@@ -383,6 +382,16 @@ void ProxyDrawOverride::draw(const MHWRender::MDrawContext& context, const MUser
 
     stateManager->setDepthStencilState(previousDepthState);
     MHWRender::MStateManager::releaseDepthStencilState(depthState);
+
+    // Check framestamp b/c we don't want to put multiple refresh commands
+    // on the idle queue for a single frame-render... especially if we have
+    // multiple ProxyShapes...
+    if (!ptr->m_engine->IsConverged() && context.getFrameStamp() != s_lastRefreshFrameStamp)
+    {
+      s_lastRefreshFrameStamp = context.getFrameStamp();
+      // Force another refresh of the current viewport
+      MGlobal::executeCommandOnIdle("refresh -cv -f");
+    }
   }
   glClearColor(clearCol[0], clearCol[1], clearCol[2], clearCol[3]);
 }