+++ usdmayautils/AL/usdmaya/utils/MeshUtils.cpp
@@ -259,6 +259,83 @@ void MeshImportContext::gatherFaceConnectsAndVertices()
       }
     }
   }
+  else
+  {
+    // check for cases where data is left handed.
+    // Maya fails
+    TfToken orientation;
+    bool leftHanded = (mesh.GetOrientationAttr().Get(&orientation, m_timeCode) && orientation == UsdGeomTokens->leftHanded);
+    if(leftHanded)
+    {
+      size_t numPoints = pointData.size();
+      size_t numFaces = faceVertexCounts.size();
+      std::vector<GfVec3f> tempNormals(numPoints, GfVec3f(0, 0, 0));
+
+      const GfVec3f* const ptemp = (const GfVec3f*)pointData.cdata();
+      GfVec3f* const pnorm = (GfVec3f*)tempNormals.data();
+      const int32_t* pcounts = (const int32_t*)faceVertexCounts.cdata();
+      const int32_t* pconnects = (const int32_t*)faceVertexIndices.cdata();
+
+      // compute each face normal, and add into the array of vertex normals.
+      for(size_t i = 0, offset = 0; i < numFaces; ++i)
+      {
+        int32_t nverts = pcounts[i];
+        const int32_t* pface =  pconnects + offset;
+
+        offset += nverts;
+
+        // grab first two points & normals, and compute edge.
+        GfVec3f v0 = ptemp[pface[0]];
+        GfVec3f v1 = ptemp[pface[1]];
+        GfVec3f n0 = pnorm[pface[0]];
+        GfVec3f n1 = pnorm[pface[1]];
+        GfVec3f n2;
+        GfVec3f e1 = v1 - v0;
+
+        // loop through each triangle in face
+        for(int32_t j = 2; j < nverts; ++j)
+        {
+          // compute last edge in tri (will have previous edge from last iteration)
+          const GfVec3f v2 = ptemp[pface[j]];
+          n2 = pnorm[pface[j]];
+          GfVec3f e2 = v2 - v0;
+
+          // compute triangle normal
+          GfVec3f fn = GfCross(e2, e1);
+          n0 += fn;
+          n1 += fn;
+          n2 += fn;
+
+          // write summed normal (with original value) back into array
+          pnorm[pface[j - 1]] = n1;
+
+          // for next iteration
+          n1 = n2;
+          e1 = e2;
+        }
+
+        // write back first and last normal
+        pnorm[pface[0]] = n0;
+        pnorm[pface[nverts-1]] = n2;
+      }
+
+      // normalise each normal in the array
+      for(int32_t j = 0; j < numPoints; ++j)
+      {
+        pnorm[j] = GfGetNormalized(pnorm[j]);
+      }
+
+      // now expand array into a set of vertex-face normals
+      {
+        normals.setLength(connects.length());
+        for(uint32_t i = 0, nf = connects.length(); i < nf; ++i)
+        {
+          int index = connects[i];
+          normals[i] = MVector(pnorm[index][0], pnorm[index][1], pnorm[index][2]);
+        }
+      }
+    }
+  }
 }
 
 //----------------------------------------------------------------------------------------------------------------------